[Sem Handler] header implementation
* local_semlock used for remote semaphore
they has checked remote effective bit number
