+++ linux-4.15/ipc/sem_handler.h
@@ -3,6 +3,12 @@
 
 #include <linux/sem.h>
 
+typedef struct semarray_object {
+	struct sem_array imported_sem;
+	struct sem_array *local_sem;
+	struct sem* mobile_sem_base;
+} semarray_object_t;
+
 int share_existing_semundo_proc_list(struct task_struct *tsk,
 				     unique_id_t undo_list_id);
 int create_semundo_proc_list(struct task_struct *tsk);
